Added more comments
