+++ integrations/azure/storage.go
@@ -4,6 +4,7 @@ import (
 	"azurekv/logger"
 	"fmt"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/security/keyvault/azkeys"
 	"github.com/keeper-security/secrets-manager-go/core"
 )
 
@@ -58,7 +59,7 @@ func (a *AzureKeyValueStorage) Delete(key core.ConfigKey) map[string]interface{}
 		logger.Debugf("Removed key: %s", string(key))
 		a.saveConfig(a.config)
 	} else {
-		logger.Warnf("%s", fmt.Sprintf("No key '%s' was found in config", string(key)))
+		logger.Warnf("No key '%s' was found in config", string(key))
 	}
 	return a.ReadStorage()
 }
@@ -76,4 +77,46 @@ func (a *AzureKeyValueStorage) IsEmpty() bool {
 func (a *AzureKeyValueStorage) Contains(key core.ConfigKey) bool {
 	_, found := a.config[key]
 	return found
+}
+
+// Changes the key used to encrypt/decrypt the configuration.
+func (s *AzureKeyValueStorage) ChangeKey(newKeyURL string) (bool, error) {
+	oldState := struct {
+		vaultURL, keyName, keyVersion string
+		cryptoClient                  *azkeys.Client
+	}{
+		s.azureConfig.KeyURL, s.keyName, s.keyVersion, s.cryptoClient,
+	}
+
+	vaultURL, keyName, keyVersion, err := fetchKeyDetails(newKeyURL)
+	if err != nil {
+		logger.Errorf("Failed to extract key details from URL '%s': %v", newKeyURL, err)
+		return false, fmt.Errorf("failed to extract key details from URL '%s': %w", newKeyURL, err)
+	}
+
+	s.azureConfig.KeyURL = newKeyURL
+	s.keyName = keyName
+	s.keyVersion = keyVersion
+
+	cred, err := fetchCredentials(s.azureConfig)
+	if err != nil {
+		return false, err
+	}
+
+	client, err := azkeys.NewClient(vaultURL, cred, nil)
+	if err != nil {
+		return false, fmt.Errorf("failed to create Azure Key Vault client: %w", err)
+	}
+
+	s.cryptoClient = client
+	if err := s.saveConfig(s.config); err != nil {
+		s.azureConfig.KeyURL = oldState.vaultURL
+		s.keyName = oldState.keyName
+		s.keyVersion = oldState.keyVersion
+		s.cryptoClient = oldState.cryptoClient
+		logger.Errorf("Failed to change the key to '%s' for config '%s': %v", newKeyURL, s.configFileLocation, err)
+		return false, fmt.Errorf("failed to change the key for %s: %w", s.configFileLocation, err)
+	}
+
+	return true, nil
 }
